refactor(tests): Split AdjacencyList unit tests per #113
